pyproject.toml
[tool.poetry]
name = "hello_world"
version = "0.1.0"
description = "CS348 Team Project"
authors = [ 
    "Alex Au <a32au@uwaterloo.ca>",
    "Anika Awasthi <a5awasth@uwaterloo.ca>",
    "Ananya Ohrie <aohrie@uwaterloo.ca>",
    "Anthony Tieu <anthony.tieu@uwaterloo.ca>",
    "Kyle Zheng <k4zheng@uwaterloo.ca>"
]
readme = "README.md"

packages = [
    { include = "DataImporter", from = "Tools/DataImporter/src" },
    { include = "PyUtils", from = "Tools/PyUtils/src" }
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
flask = "^3.1.1"
psycopg2-binary = ">=2.9.10,<3.0.0"
python-dotenv = ">=1.1.0,<2.0.0"
pandas = ">=2.3.0,<3.0.0"
sqlalchemy = ">=2.0.41,<3.0.0"
pytest = ">=8.4.0,<9.0.0"
waitress = ">=3.0.2,<4.0.0"
flask-cors = "^6.0.1"
pytz = "^2025.2"
tzlocal = "^5.3.1"

[tool.poetry.scripts]
hello_world = "hello_world.main:main"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
gunicorn = "^23.0.0"

